refactor: MatchUser MatchGame 분리 및 변경 로직 Converter로 옮김
