+++ src/main/java/com/gamemoonchul/infrastructure/web/dto/response/MatchUserResponse.java
@@ -1,0 +1,13 @@
+package com.gamemoonchul.infrastructure.web.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record MatchUserResponse(
+    Long id,
+    String nickname,
+    String championName,
+    String championThumbnail,
+    boolean win
+) {
+}
